refactor: extract cache logic into hooks
